SITE-31 feat: zap logging functionality and tests
This commit it is an addition to SITE-31. It implements zap logging and
tests. There are some features here that are worth noting:

* This zap implementation comprises of three environments: development,
staging, and production. Each logger is tested for those respective
environments. The logger can be used in all of thse environments and the
output should be as expected.

* The tests are pretty comprehensive, however some of them may be
redundant. I'll need to do more investigation. Also, there are some
parts which can be more DRY, but I'll have to refactor that later.
Overall, I'm confident in the tests.

The test suite still needs some improvements and documentation. I'll get
to that sooner and hopefully not later.

Cheers.