feat: create `pkg/errors`
Our version of `pkg/errors` unifies two error libraries: the standard
library for Go and the github.com/pkg/errors library. There is some
overlap between the two for functions, so the best performing one on
synthetic benchmarks was chosen for the implementation. There are a few
optimizations that can happen in the future for this library, however,
it does what it can at the moment.

Included are unit tests and benchmark tests. See for yourself.

refactor: move `assert` and `test` to `pkg`